+++ apps/ui/app/components/nav/sidebar/sidebar.tsx
@@ -1,21 +1,17 @@
 import { Link, useLoaderData, useLocation, useParams } from "@remix-run/react";
-import {
-  BookOpenText,
-  FlaskConical,
-  GanttChart,
-  Hammer,
-  HeartHandshake,
-  Home,
-  LayoutDashboard,
-  Rocket,
-} from "lucide-react";
+import { BookOpen, FlaskConical, GanttChart, Hammer, HelpCircle, Home, LayoutDashboard, Rocket } from "lucide-react";
 import { Routes } from "../../../routes";
 import { EnvLayoutLoader } from "../../../types/envLayout.loader.types";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "../../ui/tooltip";
+import { ComponentProps, FC } from "react";
+import { EnvSelector } from "../header/EnvSelector";
+import { ProfileDropdown } from "../header/ProfileDropdown";
+
+type ClassName = ComponentProps<"div">["className"];
 
 type BaseSidebarItem = {
   name: string;
-  icon: JSX.Element;
+  icon: FC<{ className: ClassName }>;
 }
 
 type DisabledItem = BaseSidebarItem & {
@@ -28,49 +24,55 @@ type LinkItem = BaseSidebarItem & {
 
 type SidebarItem = DisabledItem | LinkItem;
 
+type BottomSidebarItem = {
+  name: string;
+  icon: JSX.Element;
+  href: string;
+};
+
 const SidebarItems: SidebarItem[] = [
   {
     name: "Dashboard",
     href: Routes.app.project.env.dashboard,
-    icon: <LayoutDashboard size={20} />,
+    icon: ({ className }) => <LayoutDashboard size={16} className={className} />,
   },
   {
     name: "Traces & Logs",
     href: Routes.app.project.env.traces,
-    icon: <GanttChart size={20} />,
+    icon: ({ className }) => <GanttChart size={16} className={className} />,
   },
   {
     name: "Prompts & Tools",
     disabled: true,
-    icon: <Hammer size={20} />,
+    icon: ({ className }) => <Hammer size={16} className={className} />,
   },
   {
     name: "Experiments",
     disabled: true,
-    icon: <FlaskConical size={20} />,
+    icon: ({ className }) => <FlaskConical size={16} className={className} />,
   },
   {
     name: "Deployments",
     disabled: true,
-    icon: <Rocket size={20} />,
+    icon: ({ className }) => <Rocket size={16} className={className} />,
   },
 ];
 
-const BottomSidebarItems = [
+const BottomSidebarItems: BottomSidebarItem[] = [
   {
-    name: "Documentation",
+    name: "Docs",
     href: Routes.external.documentation,
-    icon: <BookOpenText size={20} />,
+    icon: <BookOpen size={16} className={"group-hover:text-foreground text-muted-foreground"} />,
   },
   {
-    name: "Support",
-    href: Routes.external.support,
-    icon: <HeartHandshake size={20} />,
+    name: "Help",
+    href: Routes.external.discord,
+    icon: <HelpCircle size={16} className={"group-hover:text-foreground text-muted-foreground"} />,
   },
 ];
 
 export const Sidebar = () => {
-  const { project } = useLoaderData<EnvLayoutLoader>();
+  const { user, project, environment } = useLoaderData<EnvLayoutLoader>();
   const { pathname } = useLocation();
   const params = useParams();
   const envId = params.envId!;
@@ -87,14 +89,14 @@ export const Sidebar = () => {
           <Tooltip>
             <TooltipTrigger asChild>
               <li key={item.name}
-                  className="flex items-center py-2 px-4 rounded-lg text-muted-foreground cursor-not-allowed">
+                  className="flex items-center py-1 cursor-not-allowed">
                 <div className="flex justify-center w-8">
-                  {item.icon}
+                  {<item.icon className={"text-muted-foreground/50"} />}
                 </div>
-                <span className="ml-2 whitespace-nowrap">{item.name}</span>
+                <span className="ml-1 whitespace-nowrap text-muted-foreground/50">{item.name}</span>
               </li>
             </TooltipTrigger>
-            <TooltipContent side={"right"} sideOffset={-40} align={"start"}>
+            <TooltipContent side={"right"} sideOffset={-10} align={"start"}>
               <p>Coming Soon</p>
             </TooltipContent>
           </Tooltip>
@@ -104,68 +106,67 @@ export const Sidebar = () => {
 
     const isActive = pathname.startsWith(item.href(params));
     return (
-      <li key={item.name}>
+      <li key={item.name} className={`${isActive ? "bg-muted/50" : ""} group hover:bg-muted/50 rounded`}>
         <Link
           to={item.href(params)}
           prefetch={"intent"}
-          className={`flex items-center py-2 px-4 rounded-lg hover:underline underline-offset-2 ${isActive ? "underline" : ""}`}
+          className={"flex items-center py-1"}
         >
-          <div className="flex justify-center w-8">
-            {item.icon}
+          <div className={"flex justify-center w-8"}>
+            {<item.icon className={"group-hover:text-foreground text-muted-foreground"} />}
           </div>
-          <span className="ml-2 whitespace-nowrap">{item.name}</span>
+          <span className="ml-1 whitespace-nowrap text-muted-foreground">{item.name}</span>
         </Link>
       </li>
     );
   });
 
   const BottomItemsComponent = () => BottomSidebarItems.map((item) => (
-    <li key={item.name}>
+    <li key={item.name} className={"group hover:bg-muted/50 rounded"}>
       <Link
         to={item.href}
         target={"_blank"}
         prefetch={"intent"}
-        className={`flex items-center py-2 px-4 rounded-lg hover:underline underline-offset-2 ${
-          pathname === item.href ? "underline" : ""
-        }`}
+        className={"flex items-center py-1"}
       >
         <div className="flex justify-center w-8">{item.icon}</div>
-        <span className="ml-2 whitespace-nowrap">{item.name}</span>
+        <span className="ml-1 whitespace-nowrap text-muted-foreground">{item.name}</span>
       </Link>
     </li>
   ));
 
   return (
-    <aside className="w-52 h-screen fixed left-0 top-0 flex flex-col" aria-label="Sidebar">
-      <div className="flex flex-row items-center py-6 px-4 gap-2">
-        <img className="h-8 w-auto" src={"/logo.png"} alt="Logo" />
-        <p className={"text-xl"}>MonteloAI</p>
+    <aside className="w-52 h-screen fixed left-0 top-0 flex flex-col border-r-[1px]" aria-label="Sidebar">
+      <div className={"flex flex-col p-4 gap-4 mb-2"}>
+        <div className="flex flex-row items-center justify-between">
+          <div className={"flex flex-row items-center"}>
+            <Link
+              to={Routes.app.root}
+              prefetch={"intent"}
+              className={"group flex items-center py-1 hover:bg-muted/50 rounded"}
+            >
+              <div className="flex justify-center w-8">
+                <Home size={16} className={"group-hover:text-foreground text-muted-foreground"} />
+              </div>
+            </Link>
+            <p className={"text-sm text-muted-foreground"}>/ {project.team.name}</p>
+          </div>
+          <ProfileDropdown user={user} />
+        </div>
+        <div className={"flex flex-col gap-1"}>
+          <p className={"text-sm text-muted-foreground"}>{project.name}</p>
+          <EnvSelector environments={project.environments} pathEnv={environment} />
+        </div>
       </div>
-      <div className="overflow-y-auto flex-grow">
-        <ul className="space-y-2">
+      <div className="overflow-y-auto flex-grow pl-4 pr-4 pb-4">
+        <ul className="space-y-1">
           <SidebarItemsComponent />
         </ul>
       </div>
-      <div className="mt-auto">
-        <ul className="space-y-2 mb-8">
+      <div className="mt-auto p-4">
+        <ul className="space-y-1">
           <BottomItemsComponent />
         </ul>
-        <div className={"flex-row items-center mb-4"}>
-          <div className={"px-4"}>
-            <p className={"text-sm text-muted-foreground font-light my-1"}>Project: {project.name}</p>
-            <p className={"text-sm text-muted-foreground font-light my-1"}>Team: {project.team.name}</p>
-          </div>
-          <Link
-            to={Routes.app.root}
-            prefetch={"intent"}
-            className={"flex items-center py-2 px-4 rounded-lg hover:underline underline-offset-2"}
-          >
-            <div className="flex justify-center w-8">
-              <Home size={20} />
-            </div>
-            <span className="ml-2 whitespace-nowrap">Home</span>
-          </Link>
-        </div>
       </div>
     </aside>
   );