+++ apps/ui/app/components/tables/LogTable/idShortener.ts
@@ -1,38 +1,34 @@
+import { BadgeProps } from "../../ui/badge";
 
-export const idShortener = (id: string): { short: string; color: string } => {
+export const idShortener = (id: string): { short: string; variant: BadgeProps["variant"] } => {
   const short = id.substring(id.length - 4);
 
   const lastLetterAscii = id.charCodeAt(id.length - 1);
-  const colors: string[] = [
-    "#E54D2E",
-    "#E5484D",
-    "#E54666",
-    "#E93D82",
-    "#D6409F",
-    "#AB4ABA",
-    "#8E4EC6",
-    "#6E56CF",
-    "#5B5BD6",
-    "#3E63DD",
-    "#0090FF",
-    "#00A2C7",
-    "#00A2C7",
-    "#12A594",
-    "#30A46C",
-    "#46A758",
-    "#F76B15",
-    "#FFC53D",
-    "#FFE629",
-    "#BDEE63",
-    "#86EAD4",
-    "#7CE2FE",
+  const variants: Array<BadgeProps["variant"]> = [
+    "pink",
+    "rose",
+    "fuchsia",
+    "purple",
+    "violet",
+    "indigo",
+    "blue",
+    "sky",
+    "cyan",
+    "teal",
+    "emerald",
+    "green",
+    "lime",
+    "yellow",
+    "amber",
+    "orange",
+    "red"
   ];
 
-  const adjustedAsciiIndex = lastLetterAscii % colors.length;
-  const color = colors[adjustedAsciiIndex];
+  const adjustedAsciiIndex = lastLetterAscii % variants.length;
+  const variant = variants[adjustedAsciiIndex];
 
   return {
     short,
-    color,
+    variant,
   };
 };