+++ apps/ui/app/routes/project+/$projectId+/env+/$envId+/_layout.tsx
@@ -2,7 +2,6 @@ import { json, LoaderFunction } from "@remix-run/node";
 import { Outlet } from "@remix-run/react";
 import { withAuth } from "../../../../../common/auth/withAuth";
 import { EnvLayoutLoader } from "../../../../../types/envLayout.loader.types";
-import { Header } from "../../../../../components/nav/header/header";
 import { Sidebar } from "../../../../../components/nav/sidebar/sidebar";
 
 export const loader: LoaderFunction = withAuth(async ({ params, api, user }) => {
@@ -26,10 +25,9 @@ export const loader: LoaderFunction = withAuth(async ({ params, api, user }) => 
 export default function DashboardLayout() {
   return (
     <div className="flex flex-col h-screen">
-      <Header />
       <div className="flex flex-1 overflow-hidden">
         <Sidebar />
-        <main className="flex-1 overflow-x-hidden overflow-y-auto pb-4 px-8 ml-48">
+        <main className="flex-1 overflow-x-hidden overflow-y-auto pb-4 px-8 ml-52 mt-8">
           <Outlet />
         </main>
       </div>