+++ apps/ui/tailwind.config.js
@@ -55,36 +55,6 @@ module.exports = {
           DEFAULT: "hsl(var(--card))",
           foreground: "hsl(var(--card-foreground))",
         },
-        purple: {
-          DEFAULT: "#6B46C1",
-          foreground: "#FFFFFF",
-          80: "#7C53C3",
-        },
-        blue: {
-          DEFAULT: "#4299E1",
-          foreground: "#FFFFFF",
-          80: "#63B3ED",
-        },
-        orange: {
-          DEFAULT: "#ED8936",
-          foreground: "#FFFFFF",
-          80: "#F6AD55",
-        },
-        green: {
-          DEFAULT: "#48BB78",
-          foreground: "#FFFFFF",
-          80: "#68D391",
-        },
-        red: {
-          DEFAULT: "#F56565",
-          foreground: "#FFFFFF",
-          80: "#FC8181",
-        },
-        yellow: {
-          DEFAULT: "#ECC94B",
-          foreground: "#2D3748",
-          80: "#F6E05E",
-        },
       },
       borderRadius: {
         lg: "var(--radius)",