+++ apps/api/CHANGELOG.md
@@ -1,5 +1,21 @@
 # actions
 
+## 1.1.41
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/db@1.1.41
+
+## 1.1.40
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/db@1.1.40
+
 ## 1.1.39
 
 ### Patch Changes