+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @montelo/cli
 
+## 1.1.41
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/promptfile@1.1.41
+  - @montelo/config@1.1.41
+  - @montelo/fizz@1.1.41
+
+## 1.1.40
+
+### Patch Changes
+
+- y
+- Updated dependencies
+  - @montelo/config@1.1.40
+  - @montelo/fizz@1.1.40
+  - @montelo/promptfile@1.1.40
+
 ## 1.1.39
 
 ### Patch Changes